seperated table component and added calendar page
